refactor(server): input form flat in one ft
